requirements.txt
tensorflow
keras
opencv-python
numpy
pillow
tk
playsound
